+++ resources/views/posts.blade.php
@@ -24,7 +24,7 @@
       @foreach ($posts as $post)
         <article>
           <h1>
-              <a href="/posts/{{$post->slug;}}">
+              <a href="/posts/{{$post->id;}}">
                 {{$post->title}} 
               </a>
           </h1>
