+++ routes/web.php
@@ -27,9 +27,9 @@ Route::get('/hello', function () {
     return view('hello');
 });
 
-Route::get('posts/{post}', function ($slug) {
+Route::get('posts/{post}', function ($id) {
     
-    $post = Post::findOrFail($slug);    
+    $post = Post::findOrFail($id);    
     return view('post', [
         'post' => $post]);
